Refactored pipeline architecture.
Given the low cost of temporary files on Linux (particularly when
created on tmpfs file systems, as is the default on many distros now), I
refactored the pipeline code to remove the previous optimization and
create temporary files whenever stdout or stderr from the overall
pipeline is set to PIPE. This way, all potentially problematic calls to
things that can misbehave in the subprocess.Popen object (particularly
communicate()) are removed. Renamed subpipe module to pipeline and
removed the convenience function for now. I plan to re-add convenience
functions later.

Added a PeekReader class, whose primary purpose is to use os.pread() to
read the contents of the temporary files without changing the file
offsets. This way, client code can monitor a long-running pipeline where
the output/error is being redirected into a temporary file, without
destroying the output data.